Somewhat functional email, images don't work locally and aren't permanent
